Move filestore-client to a separate dir
