+++ cloud/disk_manager/internal/pkg/facade/filesystem_snapshot_transfer_nemesis_test/ya.make
@@ -6,10 +6,7 @@ SET_APPEND(RECIPE_ARGS --enable-list-nodes-logging)
 SET_APPEND(RECIPE_ARGS --allow-filestore-force-destroy)
 SET_APPEND(RECIPE_ARGS --list-nodes-max-bytes=10000)
 INCLUDE(${ARCADIA_ROOT}/cloud/disk_manager/internal/pkg/facade/testcommon/common.inc)
-
-DEPENDS(
-    cloud/filestore/apps/client
-)
+INCLUDE(${ARCADIA_ROOT}/cloud/disk_manager/test/filestore_client/filestore_client.inc)
 
 GO_XTEST_SRCS(
     ../filesystem_snapshot_transfer_test/filesystem_snapshot_transfer_test.go
