+++ cloud/disk_manager/test/filestore_client/filestore_client.go
@@ -1,4 +1,4 @@
-package testcommon
+package filestore_client
 
 import (
 	"fmt"
