+++ cloud/disk_manager/test/filestore_client/filestore_client.inc
@@ -1,0 +1,7 @@
+PEERDIR(
+    cloud/disk_manager/test/filestore_client
+)
+
+DEPENDS(
+    cloud/filestore/apps/client
+)
