+++ cloud/disk_manager/test/filestore_client/ya.make
@@ -1,7 +1,7 @@
 GO_LIBRARY()
 
 SRCS(
-    common.go
+    filestore_client.go
 )
 
 END()
