+++ cloud/disk_manager/test/ya.make
@@ -1,5 +1,6 @@
 RECURSE(
     acceptance
+    filestore_client
     images
     mocks
     recipe
